requirements.txt
alembic==1.0.7
SQLAlchemy==1.2.17
mysql_connector_repackaged==0.3.1
